Renamed "committed" to "commit-reducer". Split types into separate TS files, to reduce dependencies between files. Improved docs for commit-reducer.
