+++ packages/redux/src/features/commit-reducer/state.ts
@@ -1,0 +1,4 @@
+export interface CommitReducerState<TState> {
+  workspace: TState;
+  committed: TState;
+}
