+++ packages/redux/src/index.ts
@@ -1,3 +1,5 @@
-export * from "./commited/actions";
-export * from "./commited/reducer";
-export * from "./commited/selectors";
+export * from "./features/commit-reducer/action-creators";
+export * from "./features/commit-reducer/actions";
+export * from "./features/commit-reducer/reducer";
+export * from "./features/commit-reducer/selectors";
+export * from "./features/commit-reducer/state";
